+++ ktlint-core/src/test/kotlin/com/pinterest/ktlint/core/internal/EditorConfigLoaderTest.kt
@@ -11,11 +11,12 @@ import java.nio.file.FileSystem
 import java.nio.file.Files
 import java.nio.file.Path
 import org.assertj.core.api.Assertions.assertThat
+import org.assertj.core.api.Assertions.entry
 import org.ec4j.core.model.PropertyType
 import org.intellij.lang.annotations.Language
 import org.jetbrains.kotlin.com.intellij.lang.ASTNode
-import org.junit.After
-import org.junit.Test
+import org.junit.jupiter.api.AfterEach
+import org.junit.jupiter.api.Test
 
 @OptIn(FeatureInAlphaState::class)
 internal class EditorConfigLoaderTest {
@@ -36,7 +37,7 @@ internal class EditorConfigLoaderTest {
         Files.write(normalizedPath("$filePath/.editorconfig"), content.toByteArray())
     }
 
-    @After
+    @AfterEach
     fun tearDown() {
         tempFileSystem.close()
     }
@@ -240,21 +241,33 @@ internal class EditorConfigLoaderTest {
     }
 
     @Test
-    fun `Should return emtpy map on null file path`() {
-        val parsedEditorConfig = editorConfigLoader.loadPropertiesForFile(null, rules = rules)
+    fun `Should return the override properties only on null file path`() {
+        val parsedEditorConfig = editorConfigLoader.loadPropertiesForFile(
+            filePath = null,
+            rules = rules,
+            loadedValuesOverride = mapOf(
+                FinalNewlineRule.insertNewLineProperty to PropertyType.PropertyValue.valid("true", true)
+            )
+        )
 
-        assertThat(parsedEditorConfig).isEmpty()
+        assertThat(parsedEditorConfig.convertToRawValues()).containsExactly(
+            entry("insert_final_newline", "true")
+        )
     }
 
     @Test
-    fun `Should return empty map for non supported file`() {
-        val projectDir = "/project"
-        val lintFile = tempFileSystem.normalizedPath(projectDir).resolve("test.txt")
+    fun `Should return the override properties only non supported file`() {
+        val parsedEditorConfig = editorConfigLoader.loadPropertiesForFile(
+            filePath = null,
+            rules = rules,
+            loadedValuesOverride = mapOf(
+                FinalNewlineRule.insertNewLineProperty to PropertyType.PropertyValue.valid("true", true)
+            )
+        )
 
-        val editorConfigProperties = editorConfigLoader.loadPropertiesForFile(lintFile, rules = rules)
-        val parsedEditorConfig = editorConfigProperties.convertToRawValues()
-
-        assertThat(parsedEditorConfig).isEmpty()
+        assertThat(parsedEditorConfig.convertToRawValues()).containsExactly(
+            entry("insert_final_newline", "true")
+        )
     }
 
     @Test
@@ -485,5 +498,13 @@ internal class EditorConfigLoaderTest {
         ) {
             throw NotImplementedError()
         }
+    }
+
+    private companion object {
+        val SOME_EDITOR_CONFIG_PROPERTY = FinalNewlineRule.insertNewLineProperty
+        const val SOME_EDITOR_CONFIG_PROPERTY_VALUE = true
+        val SOME_EDITOR_CONFIG_OVERRIDE = mapOf(
+            SOME_EDITOR_CONFIG_PROPERTY to PropertyType.PropertyValue.valid("true", SOME_EDITOR_CONFIG_PROPERTY_VALUE)
+        )
     }
 }