+++ ktlint-reporter-baseline/src/main/kotlin/com/pinterest/ktlint/reporter/baseline/BaselineReporterProvider.kt
@@ -1,10 +1,9 @@
 package com.pinterest.ktlint.reporter.baseline
 
-import com.pinterest.ktlint.core.Reporter
 import com.pinterest.ktlint.core.ReporterProvider
 import java.io.PrintStream
 
-public class BaselineReporterProvider : ReporterProvider {
+public class BaselineReporterProvider : ReporterProvider<BaselineReporter> {
     override val id: String = "baseline"
-    override fun get(out: PrintStream, opt: Map<String, String>): Reporter = BaselineReporter(out)
+    override fun get(out: PrintStream, opt: Map<String, String>): BaselineReporter = BaselineReporter(out)
 }