+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/NoMultipleSpacesRule.kt
@@ -7,23 +7,30 @@ import org.jetbrains.kotlin.com.intellij.lang.ASTNode
 import org.jetbrains.kotlin.com.intellij.psi.PsiWhiteSpace
 import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.LeafPsiElement
 
-class NoMultipleSpacesRule : Rule("no-multi-spaces") {
-
+public class NoMultipleSpacesRule : Rule("no-multi-spaces") {
     override fun visit(
         node: ASTNode,
         autoCorrect: Boolean,
         emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit
     ) {
-        if (node is PsiWhiteSpace && !node.textContains('\n') && node.getTextLength() > 1 &&
-            // allow multiple spaces in KDoc in case of KDOC_TAG for alignment, e.g.
-            // @param foo      stuff
-            // @param foobar   stuff2
-            !(node.treePrev?.elementType == KDOC_MARKDOWN_LINK && node.treeParent?.elementType == KDOC_TAG)
-        ) {
-            emit(node.startOffset + 1, "Unnecessary space(s)", true)
-            if (autoCorrect) {
-                (node as LeafPsiElement).rawReplaceWithText(" ")
+        node
+            .takeIf { node is PsiWhiteSpace }
+            .takeUnless { node.isIndentation() }
+            .takeIf { node.textLength > 1 }
+            .takeUnless { node.isPossibleAlignmentOfKdocTag() }
+            ?.let {
+                emit(node.startOffset + 1, "Unnecessary long whitespace", true)
+                if (autoCorrect) {
+                    (node as LeafPsiElement).rawReplaceWithText(" ")
+                }
             }
-        }
     }
+
+    private fun ASTNode.isIndentation() = this.text.startsWith("\n")
+
+    // allow multiple spaces in KDoc in case of KDOC_TAG for alignment, e.g.
+    // @param foo      stuff
+    // @param foobar   stuff2
+    private fun ASTNode.isPossibleAlignmentOfKdocTag() =
+        treePrev?.elementType == KDOC_MARKDOWN_LINK && treeParent?.elementType == KDOC_TAG
 }